use single compose file & move state to internal
