--- cmd/nigiri/resources/docker-compose-regtest.yml
@@ -1,82 +1,0 @@
-version: '3.5'
-services:
-  # RPC daemon
-  bitcoin:
-    image: ghcr.io/vulpemventures/bitcoin:latest
-    container_name: bitcoin
-    command:
-      - -datadir=config
-    ports:
-      - 18443:18433
-      - 18444:18444
-    volumes:
-      - ./volumes/bitcoin/:/config
-    restart: unless-stopped
-
-  # Block explorer server
-  electrs:
-    image: ghcr.io/vulpemventures/electrs:latest
-    container_name: electrs
-    entrypoint:
-      - /build/electrs
-    command:
-      - -vvvv
-      - --network
-      - regtest
-      - --daemon-dir
-      - /config
-      - --daemon-rpc-addr
-      - bitcoin:18443
-      - --cookie
-      - admin1:123
-      - --http-addr
-      - 0.0.0.0:30000
-      - --electrum-rpc-addr
-      - 0.0.0.0:50000
-      - --cors
-      - "*"
-    depends_on:
-      - bitcoin
-    ports:
-      - 50000:50000
-      - 30000:30000
-    volumes:
-      - ./volumes/bitcoin/:/config
-    restart: unless-stopped
-
-  # Block explorer frontend
-  esplora:
-    image: ghcr.io/vulpemventures/esplora:latest
-    container_name: esplora
-    depends_on:
-      - chopsticks
-    environment: 
-        API_URL: http://localhost:3000
-    ports:
-      - 5000:5000
-    restart: unless-stopped
-
-  # Chopsticks
-  chopsticks:
-    image: ghcr.io/vulpemventures/nigiri-chopsticks:latest
-    container_name: chopsticks
-    command:
-      - --use-faucet
-      - --use-mining
-      - --use-logger
-      - --rpc-addr
-      - bitcoin:18443
-      - --electrs-addr
-      - electrs:30000
-      - --addr
-      - 0.0.0.0:3000
-    depends_on:
-      - bitcoin
-      - electrs
-    ports:
-      - 3000:3000
-    restart: unless-stopped
-
-networks:
-  default:
-    name: nigiri
